Added experimental support for Mercurial
Implemented a new HistorReader for Mercurial support. Due to the fact
that we don't have a Java binding for Mercurial yet, this prototype
will wrap around the shell commands. This gave a poor perfomance for
the history reader, so it will cache the output from the hg log
commands into a file. The web-server must set "SCAN_REPOS" to true in
web.xml to let the webservice scan for Mercurial repositories during
startup. The path to the repositories must be specified with the -m
option when the index database should be created.