This fixes the following list.jsp bug with annotating Subversion repositories.
The problem is that the revisions in 'blame' may go back further in history
than the currently existing path, especially when the file is copied.  When
the user clicks on one of these links, an error message is displayed.

This patch finds the oldest revision with stop-on-copy on this path and uses
the revision as a criteria for disabling revision links that will cause the
error.