Remove scope GUI
There are a lot of bugs in the GUI, and it is not being used. It is easier to
delete it than to spend time fixing it (since noone use it). We can allways
get it back from the SCM if someone wants it back..

--HG--
rename : src/org/opensolaris/opengrok/search/scope/SearchEngine.java => src/org/opensolaris/opengrok/search/SearchEngine.java