Bug #3624: History cache should be stored in Java DB
Added a first increment of the new history cache. JDBCHistoryCache has
the same interface as FileHistoryCache and can be used as a drop-in
replacement. The current implementation is not optimized (for instance
no pooling of connections or prepared statements) and is only intended
for testing.

To enable and test JDBCHistoryCache, set the property
org.opensolaris.opengrok.useJdbcCache to true.